refactoring distributions and their containers, refs #2925, #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@23341 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netgen/NGNode.h
@@ -100,7 +100,7 @@ public:
      *
      * @return The maximum neighbour number of the node
      */
-    double getMaxNeighbours() {
+    int getMaxNeighbours() {
         return myMaxNeighbours;
     }
 
@@ -109,7 +109,7 @@ public:
      *
      * @param[in] value The new maximum neighbour number of the node
      */
-    void setMaxNeighbours(double value) {
+    void setMaxNeighbours(int value) {
         myMaxNeighbours = value;
     }
 
@@ -201,7 +201,7 @@ private:
     Position myPosition;
 
     /// @brief The maximum number of neighbours
-    double myMaxNeighbours;
+    int myMaxNeighbours;
 
     /// @brief Information whether this is the center of a cpider-net
     bool myAmCenter;
